dnsdist: Move carbon configuration into a separate header file
